// legacy/20240613-components-question-PasteForm/types.ts
export interface IOption {
  id: string;
  content: string;
}

export interface IParsedQuestion {
  id: string;
  number: number;
  content: string;
  options: IOption[];
  answer: number | null;
  tags: ITag[];
  images: { url: string; hash: string }[];
  explanationImages: { url: string; hash: string }[];
  examples?: string[];
  explanation?: string;
  created_at?: string;
  updated_at?: string;
}

export interface IResponseData {
  success: boolean;
  message: string;
  questions?: IParsedQuestion[];
}

export interface ITag {
  id: string;
  name: string;
  color: string;
}